refactor(project): own evaluator scaffold types in the handler layer (DI)
Addresses PR feedback: the handler was importing EVALUATOR_LIBRARIES +
ManagedEvaluatorScaffoldInput from core/project/templates, coupling the UI
layer to core. Mirror runtime's ScaffoldRuntimeInput placement: define
EVALUATOR_LIBRARIES (names) + EvaluatorLibrary + ManagedEvaluatorScaffoldInput
in handlers/project/types.ts; templates/evaluator.ts imports those and keeps
the asset/timeout map (EVALUATOR_ASSETS). Arrow now points core -> handler-types.
